+++ mpr/eval.py
@@ -1,23 +1,15 @@
 import argparse
-from collections import defaultdict
+import torch
 import datetime
 import os
 import os.path as osp
 import random
-
-import networkx as nx
 import numpy as np
+
 from sklearn.model_selection import StratifiedKFold
-import torch
-from torch.nn import Linear, ModuleList, BatchNorm1d, Sequential
-import torch.nn.functional as F
-from torch_geometric.nn import GCNConv, GINConv
 from torch_geometric.datasets import TUDataset
 from torch_geometric.transforms import OneHotDegree
-from torch_geometric.utils import to_networkx, dense_to_sparse
-from torch.utils.data import Subset
 
-from utils import dense_diff_pool, dense_mincut_pool, to_dense_adj
 
 EPS = 1e-5
 
@@ -58,242 +50,6 @@ log_name = '%s_%s_%d_%s_%d_%.4f_%.2f' % (str(args.cluster_dims),
                                          args.lrate_anneal_coef)
 
 
-def generate_intervals(num=5, overlap=0.10, vmin=0, vmax=1):
-    """Generates the Mapper intervals in [0, 1] with the specified overlap.
-
-    params:
-      num: Number of intevals
-      overlap: Percentage specifying overlap between consecutive segments.
-    """
-    assert 0.0 <= overlap <= 1.0 # Overlap must be a percentage
-
-    overlap *= 1. / (num + 1)
-    mid = np.linspace(vmin, vmax, num+1)
-    high = mid + overlap
-    low = mid - overlap
-    intervals = np.concatenate((low[:-1, None], high[1:, None]), axis=1)
-    return intervals
-
-
-def compute_preimages(G, f, intervals=generate_intervals()):
-    """Computes the preimages of the function f using the specified intervals."""
-    preimages = []
-    for interv in intervals:
-        preimages.append(np.arange(G.number_of_nodes())[np.logical_and(
-            interv[0] <= f, f <= interv[1])])
-    return preimages
-
-
-def build_mapper_graph(G, preimages, cluster=False, return_nx=False):
-    """Given a graph G and the preimages of the function f on G it builds the
-      mapper graph.
-
-    params:
-      G: The graph to be visualized.
-      preimages: The preimages of a function f on the nodes of graph G.
-    """
-    mnode = 0
-    mnode_to_nodes = []
-    mnode_to_color = []
-    edge_weight = defaultdict(int)
-    current_adj = torch.tensor(
-        nx.to_numpy_matrix(G, nodelist=np.arange(G.number_of_nodes()),
-                           dtype=np.float32)).to(device)
-
-    # Each preimage has a corresponding colour given by its index.
-    for color, pre_img in enumerate(preimages):
-        # Skip if the preimage is empty
-        if not len(pre_img):
-            continue
-
-        # Build the subgraph of the preimage and determine the connected components.
-        if cluster:
-            G_pre = G.subgraph(pre_img)
-            connected_components = list(nx.connected_components(G_pre))
-        else:
-            connected_components = [pre_img]
-
-        for cc in connected_components:
-            # Make each connected component a node and assign its color.
-            mnode_to_node = torch.zeros(G.number_of_nodes()).to(device)
-            mnode_to_node[np.fromiter(cc, int, len(cc))] = 1.0
-            mnode_to_nodes.append(mnode_to_node)
-            mnode_to_color.append(color)
-            mnode += 1
-
-    # Initialise Mapper graph
-    mnode_to_nodes = torch.stack(mnode_to_nodes)
-    mnode_to_nodes = mnode_to_nodes / torch.sum(mnode_to_node, dim=0)
-    adj = mnode_to_nodes.mm(current_adj).mm(mnode_to_nodes.t())
-
-    if return_nx:
-        MG = nx.from_numpy_matrix(adj.cpu().numpy())
-    else:
-        MG = None
-
-    return MG, [mnode_to_nodes.cpu().numpy(),
-                mnode_to_color,
-                mnode_to_nodes.t().cpu().numpy(),
-                adj]
-
-
-class GEmbedNet(torch.nn.Module):
-    def __init__(self, input_dim=89, hidden_dim=128):
-        super(GEmbedNet, self).__init__()
-        self.conv1 = GCNConv(input_dim, hidden_dim)
-        self.conv2 = GCNConv(hidden_dim, hidden_dim)
-
-    def forward(self, x, edge_index):
-        x = F.relu(self.conv1(x, edge_index))
-        x = F.relu(self.conv2(x, edge_index))
-        return x
-
-
-class GClassifier(torch.nn.Module):
-    def __init__(self, input_dim=128, hidden_dim=128):
-        super(GClassifier, self).__init__()
-        self.conv1 = GCNConv(input_dim, hidden_dim)
-        self.conv2 = GCNConv(hidden_dim, hidden_dim)
-        self.mlp = Linear(hidden_dim, dataset.num_classes)
-
-    def forward(self, x, edge_index, edge_attr):
-        x = F.relu(self.conv1(x, edge_index, edge_attr))
-        x = F.relu(self.conv2(x, edge_index, edge_attr))
-
-        x = F.dropout(x, training = self.training)
-        x = torch.mean(x, dim=0)
-        x = self.mlp(x)
-        return x
-
-
-class StandardPoolingModel(torch.nn.Module):
-    def __init__(self, mode='diffpool', hidden_dim=32):
-        assert mode in ['diffpool', 'mincut']
-        super(StandardPoolingModel, self).__init__()
-        self.graph_convs = ModuleList([
-            GCNConv(dataset.num_node_features, hidden_dim),
-            GCNConv(hidden_dim, hidden_dim),
-            GCNConv(hidden_dim, hidden_dim)])
-        self.graph_skips = ModuleList([
-            Linear(dataset.num_node_features, hidden_dim),
-            Linear(hidden_dim, hidden_dim),
-            Linear(hidden_dim, hidden_dim)])
-
-        max_num_nodes = np.max(dataset.data.num_nodes)
-        pool_dims = [int(max_num_nodes * args.pooling_ratio),
-                     int(max_num_nodes * args.pooling_ratio**2)]
-        self.mode = mode
-        self.pooling_fn = (dense_diff_pool if mode == 'diffpool'
-                           else dense_mincut_pool)
-        if mode == 'mincut':
-            self.assignment_ws = ModuleList([
-                Linear(dataset.num_node_features, hidden_dim),
-                Linear(hidden_dim, pool_dims[0]),
-                Linear(hidden_dim, hidden_dim),
-                Linear(hidden_dim, pool_dims[1])])
-        else:
-            self.pool_convs = ModuleList([
-                GCNConv(dataset.num_node_features, pool_dims[0]),
-                GCNConv(hidden_dim, pool_dims[1])])
-
-        self.classifier = Linear(hidden_dim, dataset.num_classes)
-
-    def forward(self, x, edge_index):
-        total_loss1 = 0
-        total_loss2 = 0
-        edge_attr = None
-
-        for i in range(len(self.graph_convs)):
-            if i < len(self.graph_convs) - 1:
-                if self.mode == 'mincut':
-                    s = self.assignment_ws[2*i+1](F.relu(
-                        self.assignment_ws[2*i](x)))
-                else:
-                    s = self.pool_convs[i](x, edge_index, edge_attr)
-
-            x = F.relu((self.graph_convs[i](x, edge_index, edge_attr) +
-                        self.graph_skips[i](x)))
-
-            if i < len(self.graph_convs) - 1:
-                x, adj, loss1, loss2 = self.pooling_fn(
-                    x,
-                    to_dense_adj(edge_index, x.size(0), edge_attr=edge_attr),
-                    s)
-                edge_index, edge_attr = dense_to_sparse(adj.squeeze(0))
-                x = x.squeeze(0)
-                total_loss1 += loss1
-                total_loss2 += loss2
-
-        x_avg = torch.mean(x, dim=0).unsqueeze(0)
-        out = self.classifier(x_avg)
-
-        return out, total_loss1, total_loss2
-
-
-class FlatModel(torch.nn.Module):
-    def __init__(self, hidden_dim=32):
-        super(FlatModel, self).__init__()
-        self.graph_convs = ModuleList([
-            GCNConv(dataset.num_node_features, hidden_dim),
-            GCNConv(hidden_dim, hidden_dim)])
-        self.graph_skips = ModuleList([
-            Linear(dataset.num_node_features, hidden_dim),
-            Linear(hidden_dim, hidden_dim)])
-
-        self.classifier = Linear(hidden_dim, dataset.num_classes)
-
-    def forward(self, x, edge_index):
-        for i in range(len(self.graph_convs)):
-            x = F.relu((self.graph_convs[i](x, edge_index) +
-                        self.graph_skips[i](x)))
-
-        x_avg = torch.mean(x, dim=0).unsqueeze(0)
-        out = self.classifier(x_avg)
-
-        return out
-
-
-class GINModel(torch.nn.Module):
-    def __init__(self, hidden_dim=64):
-        super(GINModel, self).__init__()
-        self.gin_layers = ModuleList([])
-        input_dim = dataset.num_node_features
-        for i in range(5):
-            mlp = Sequential(
-                Linear(input_dim, hidden_dim),
-                BatchNorm1d(hidden_dim),
-                Linear(hidden_dim, hidden_dim),
-                BatchNorm1d(hidden_dim)
-            )
-            self.gin_layers.append(GINConv(nn=mlp))
-            input_dim = hidden_dim
-
-        self.classifier = Linear(5 * hidden_dim, dataset.num_classes)
-
-    def forward(self, x, edge_index):
-        reprs = []
-        for i in range(5):
-            x = self.gin_layers[i](x, edge_index)
-            reprs.append(torch.sum(x, dim=0))
-
-        summary = torch.cat(reprs, dim=0).unsqueeze(0)
-        out = self.classifier(summary)
-
-        return out
-
-
-class AverageMLP(torch.nn.Module):
-    def __init__(self):
-        super(AverageMLP, self).__init__()
-        self.classifier = Linear(dataset.num_node_features, dataset.num_classes)
-
-    def forward(self, x, edge_index):
-        summary = torch.mean(x, dim=0).unsqueeze(0)
-        out = self.classifier(summary)
-
-        return out
-
-
 def get_graph_classification_dataset(dataset):
     node_transform = None
     if dataset in ['COLLAB', 'REDDIT-BINARY', 'IMDB-BINARY', 'IMDB-MULTI',
@@ -315,64 +71,6 @@ def update_lrate(optimizer, epoch):
     if args.lrate_anneal_coef and epoch >= args.epochs // 2:
         optimizer.param_groups[0]['lr'] = (optimizer.param_groups[0]['lr'] *
                                            args.lrate_anneal_coef)
-
-
-def mpr_forward(data, pool=True):
-    # Get node embeddings f(G)
-    data = data.to(device)
-    max_num_nodes = np.max(dataset.data.num_nodes)
-
-    data.edge_attr = None
-    for i in range(len(args.hidden_dims)):
-        # Get node embeddings (first part of the lens)
-        x, edge_index = data.x, data.edge_index
-        x = g_embed_nets[i](x, edge_index)
-        vv = np.zeros(len(x))
-
-        if pool:
-            # Compute the PageRank function of node embeddings (second part of lens)
-            G = nx.to_undirected(to_networkx(data))
-            r_dict = nx.pagerank_scipy(G)
-            r = np.zeros(G.number_of_nodes())
-
-            for node, rval in r_dict.items():
-                r[node] = rval
-            ff = r
-            ff -= np.min(ff)
-            ff /= max(np.max(ff), EPS)
-            ff = np.ravel(ff)
-            vv = ff
-
-            # Generate intervals
-            curr_cluster_dim = args.cluster_dims[i]
-            intervals = generate_intervals(num=curr_cluster_dim,
-                                           overlap=args.interval_overlap)
-
-            # Compute pre-image f^-1(G)
-            preimages = compute_preimages(G, vv, intervals)
-            _, [mnode_to_nodes, mnode_to_color, node_to_mnode, adj] = build_mapper_graph(
-                G, preimages, cluster=False, return_nx=False)
-
-            mnode_features = torch.empty(size=(len(mnode_to_color), x.size(1)),
-                                         dtype=torch.float).to(device)
-
-            for mn in range(len(mnode_to_color)):
-                mnode_features[mn] = torch.sum(x[torch.BoolTensor(mnode_to_nodes[mn])],
-                                               dim=0)
-
-            x = mnode_features
-            edge_index, edge_attr = dense_to_sparse(adj)
-
-        # Update data for new embedding/pooling step
-        data.x = x
-        data.num_nodes = x.size(0)
-        data.edge_index = edge_index
-        data.edge_attr = edge_attr
-
-    # Classify resulting graph
-    y_pred = g_classifier(x, edge_index, edge_attr)
-
-    return vv, y_pred
 
 
 def select_subset(this_dataset, indices):
