+++ requirements.txt
@@ -4,10 +4,7 @@ numpy==1.18.1
 pandas==0.24.2
 scikit-learn==0.22.1
 scipy==1.4.1
-torch-cluster==1.4.5
-torch-scatter==1.4.0
-torch-sparse==0.4.3
-torch-geometric==1.3.2
 pydot==1.4.1
 graphviz==0.13.2
 umap-learn==0.4.1
+grakel==0.1b7
